# www/openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"servers":[{"url":"/api/v1"}],"paths":{"/notebooks":{"get":{"tags":["notebooks"],"summary":"List Notebooks","operationId":"list_notebooks_notebooks_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotebooksGetResponse"}}}}},"security":[{"HTTPBearer":[]}]},"post":{"tags":["notebooks"],"summary":"Create Notebook","operationId":"create_notebook_notebooks_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotebookPostRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotebookPostResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/notebooks/{notebook_id}/contexts":{"get":{"tags":["notebooks","contexts"],"summary":"List Contexts","operationId":"list_contexts_notebooks__notebook_id__contexts_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"notebook_id","in":"path","required":true,"schema":{"type":"string","title":"Notebook Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ContextsGetResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/notebooks/{notebook_id}/contexts/metadata":{"post":{"tags":["notebooks","contexts"],"summary":"Upload Metadata","operationId":"upload_metadata_notebooks__notebook_id__contexts_metadata_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"notebook_id","in":"path","required":true,"schema":{"type":"string","title":"Notebook Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ContextMetadataPostRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ContextMetadataPostResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/notebooks/{notebook_id}/contexts/{context_id}/upload":{"post":{"tags":["notebooks","contexts"],"summary":"Upload File","operationId":"upload_file_notebooks__notebook_id__contexts__context_id__upload_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"context_id","in":"path","required":true,"schema":{"type":"string","title":"Context Id"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_file_notebooks__notebook_id__contexts__context_id__upload_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Body_upload_file_notebooks__notebook_id__contexts__context_id__upload_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_file_notebooks__notebook_id__contexts__context_id__upload_post"},"ContextGetResponse":{"properties":{"fileUrl":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Fileurl","description":"The URL of the file"},"id":{"type":"string","title":"Id","description":"The unique identifier of the context"},"name":{"type":"string","title":"Name","description":"The name of the context"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description","description":"The description of the context"},"type":{"$ref":"#/components/schemas/ContextType","description":"The type of the context, e.g., PDF, text"}},"type":"object","required":["id","name","type"],"title":"ContextGetResponse"},"ContextMetadataPostRequest":{"properties":{"name":{"type":"string","title":"Name","description":"The name of the context"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description","description":"The description of the context"},"type":{"$ref":"#/components/schemas/ContextType","description":"The type of the context"}},"type":"object","required":["name","type"],"title":"ContextMetadataPostRequest"},"ContextMetadataPostResponse":{"properties":{"name":{"type":"string","title":"Name","description":"The name of the context"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description","description":"The description of the context"},"type":{"$ref":"#/components/schemas/ContextType","description":"The type of the context, e.g., PDF, text"},"id":{"type":"string","title":"Id","description":"The unique identifier of the context"}},"type":"object","required":["name","type","id"],"title":"ContextMetadataPostResponse"},"ContextType":{"type":"string","enum":["pdf","external_url","free_form_text"],"title":"ContextType"},"ContextsGetResponse":{"properties":{"contexts":{"items":{"$ref":"#/components/schemas/ContextGetResponse"},"type":"array","title":"Contexts","description":"The list of contexts"}},"type":"object","required":["contexts"],"title":"ContextsGetResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"NotebookGetResponse":{"properties":{"name":{"type":"string","title":"Name","description":"The name of the notebook"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description","description":"The description of the notebook"},"id":{"type":"string","title":"Id","description":"The unique identifier of the notebook"},"createdAt":{"type":"string","format":"date-time","title":"Createdat","description":"The date and time the notebook was created"},"updatedAt":{"type":"string","format":"date-time","title":"Updatedat","description":"The date and time the notebook was last updated"},"ownerId":{"type":"string","title":"Ownerid","description":"The unique identifier of the user who owns the notebook"}},"type":"object","required":["name","id","createdAt","updatedAt","ownerId"],"title":"NotebookGetResponse"},"NotebookPostRequest":{"properties":{"name":{"type":"string","title":"Name","description":"The name of the notebook"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description","description":"The description of the notebook"}},"type":"object","required":["name"],"title":"NotebookPostRequest"},"NotebookPostResponse":{"properties":{"name":{"type":"string","title":"Name","description":"The name of the notebook"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description","description":"The description of the notebook"},"id":{"type":"string","title":"Id","description":"The unique identifier of the notebook"},"createdAt":{"type":"string","format":"date-time","title":"Createdat","description":"The date and time the notebook was created"},"updatedAt":{"type":"string","format":"date-time","title":"Updatedat","description":"The date and time the notebook was last updated"}},"type":"object","required":["name","id","createdAt","updatedAt"],"title":"NotebookPostResponse"},"NotebooksGetResponse":{"properties":{"notebooks":{"items":{"$ref":"#/components/schemas/NotebookGetResponse"},"type":"array","title":"Notebooks","description":"The list of notebooks"}},"type":"object","required":["notebooks"],"title":"NotebooksGetResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}